feat: 취향 등록 페이지 분류
